--HG--
branch : week-of-code

+++ src/enzo/StarBuffer.h
@@ -22,8 +22,8 @@ struct StarBuffer {
   float		accretion_rate[MAX_ACCR];
   float         last_accretion_rate;
   FLOAT	accretion_time[MAX_ACCR];
-  float		Mass;
-  float		FinalMass;
+  double       	Mass;
+  double       	FinalMass;
   float		DeltaMass;
   float		BirthTime;
   float		LifeTime;
